Update player identification and correct typos in file names
The in test descriptions, the identification used for players has been revised from "a player with a RSI Handle of {0}" to "a player named {0}", making it broader and more general in scope. In addition, typographical errors in the file names 'CaptainDispandACrewParty.feature.cs' and 'CaptainDispandACrewParty.feature' have been corrected, they are now 'CaptainDisbandACrewParty.feature.cs' and 'CaptainDisbandACrewParty.feature' respectively.