Refactor CrewParty creation and update PlayerDocument model
Several refactors have been done on the CrewParty creation to remove the DateTime parameter making it more simplified and easier to use. This commit also includes the creation of a new PlayerDocument model which makes it easier to manage the player's information. Furthermore, changes were made to the UpdateMembers method to use the newly created PlayerDocument model.