Refactor startup services and updating functions
In the Startup.cs, services registration has been refactored to extract the creation of KeySecretsProvider into a separate method. In CreateCrewFunction, the route and Function class have been updated and error handling has been added. A reference to the CrewFinding project has also been added to the API project file.