Refactor crew and player classes, add new tests
Rearranged and renamed classes related to crews and players for better project structure and readablility, with CosmosDb and repositories moved to their respective folders. Updated test files accordingly. Also added player query for checking if a player is already in a crew and implemented this in various methods, replacing the previous crew query.