Implement GetPlayer method in PlayerQueries
This commit introduces the GetPlayer method in PlayerQueries to retrieve player data. A corresponding test, Get_Player_Successfully, is added to validate the functionality. In the PlayerDocument class, a ToPlayer helper method is implemented for document-to-player object transformation. Additionally, PlayerQueries is now properly wired up in the test Startup.cs.