Update crew tests and handle new exceptions
The Crew tests have been updated with additional methods and alterations to create more specific cases, including tests for when a player is a captain in an active crew. In addition, the handling of new exceptions such as LanguageNameLengthException has been incorporated into the CreateCrewFunction. Changes have also been made to the CrewBuilder and CrewQueryRepository for better test management.