Refactor code by renaming namespaces and fixing property name in Crew.cs
This commit mainly renames namespaces for clarity and organises the structure of the code base. Specifically, 'CreateRequests', 'JoinRequests', and 'Disbands' are moved to 'Commands' directory, while 'RecentCrewsRetrievalTest' is moved to 'Queries'. At the same time, the 'CreationAt' property in Crew.cs is renamed to 'CreatedAt' for consistency. Some mock classes are updated as well to reflect these changes.