Refactor CrewJoiner to use UserSession for player identification
Changes have been made to the CrewJoiner class to replace the explicit passing of player ID with a UserSession instance for identifying the current user. This change enhances code readability and consistency across the system. Furthermore, some static helper methods and constructors were added to facilitate the development of unit tests.