Refactor user's function to leave crew
This commit includes adjustments for the functionality where a player leaves a crew. Instead of identifying the crew explicitly via a crewId, the updated logic now uses the session's UserId to identify the player's active crew and remove the player from it. Some exception handling for non-existent crews and the testing corresponding to these cases has also been removed.