Add kick member from crew feature tests
Implemented feature tests for the new ability of a captain to kick members from their crew. Also, included checks for error scenarios such as unsuccessful kick attempts when a player is not a captain or when trying to kick a player not in the crew. Adjustments in related test classes were made to accommodate these new tests.