Move PlayerCollection to Players namespace
The PlayerCollection class has been moved from the Crews namespace to the Players namespace, making it more semantically correct. A using statement for Players namespace has been added in CrewResponse model for this adjustment. In CrewQueryResultContext class, the Crews property is also now initialized by default to an empty array.