Add VoiceChannelId to CrewDocument and corresponding tests
A new field, VoiceChannelId, has been added to the CrewDocument in CosmosDB. Appropriate test cases to ensure the functionality of the new field are also implemented. At the same time, SetVoiceChannel method in the IVoicedCrewCommandRepository interface and its related instances have been renamed to AddVoiceChannel to reflect its true functionality.