Refactor startup services and add Discord integration
This update removes the RepositoryService and adds repository setup to a new InfrastructureServices class. A Discord integration is added to the startup services and is set up with the required dependencies. Also, the ContainerBuilder is renamed to CosmosDbContainerBuilder and refactored for better usage. Changes are also reflected in the startup code and settingsjson file.