Add creation timestamp to VoiceChannelDocument
The field 'CreateAt' has been added to the VoiceChannelDocument class to record the creation date and time of each new voice channel. These changes are reflected in the VoicedCrewCommandRepository and its respective tests. The 'crewId' in the test is also changed from a hardcoded value to a new GUID for each run.