Rename "ExpireChannelRemover" to "ExpiredChannelRemover" and add functionality
The classes and interfaces related to the removal of expired channels were renamed to improve the clarity of their purpose. An additional functionality was also added to the 'ExpiredChannelRemover' to be able to retrieve and handle expired channels. Furthermore, a new file was created to test the removal function of expired voice channels.